Update entrypoint.sh (#1021)
* Update entrypoint.sh

This adds a line to run a `pip install -U` on setuptools and pip itself. This should update them before bundling up the built Python version.

This should help to prevent scanners from flagging setuptools and pip as vulnerable to CVEs, because the versions that come from upstream are quite old.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>